// frontend-tsx/src/_state/actions-types/types.tsx
export enum ActionType {
    LOGIN_SUCCESS = 'LOGIN_SUCCESS',
    LOGIN_FAIL = 'LOGIN_FAIL',
    SIGNUP_SUCCESS = 'SIGNUP_SUCCESS',
    SIGNUP_FAIL = 'SIGNUP_FAIL',
    ACTIVATION_SUCCESS = 'ACTIVATION_SUCCESS',
    ACTIVATION_FAIL = 'ACTIVATION_FAIL',
    USER_LOADED_SUCCESS = 'USER_LOADED_SUCCESS',
    USER_LOADED_FAIL = 'USER_LOADED_FAIL',
    AUTHENTICATED_SUCCESS = 'AUTHENTICATED_SUCCESS',
    AUTHENTICATED_FAIL = 'AUTHENTICATED_FAIL',
    PASSWORD_RESET_SUCCESS = 'PASSWORD_RESET_SUCCESS',
    PASSWORD_RESET_FAIL = 'PASSWORD_RESET_FAIL',
    PASSWORD_RESET_CONFIRM_SUCCESS = 'PASSWORD_RESET_CONFIRM_SUCCESS',
    PASSWORD_RESET_CONFIRM_FAIL = 'PASSWORD_RESET_CONFIRM_FAIL',
    GOOGLE_AUTH_SUCCESS = 'GOOGLE_AUTH_SUCCESS',
    GOOGLE_AUTH_FAIL = 'GOOGLE_AUTH_FAIL',
    GITHUB_AUTH_SUCCESS = 'GITHUB_AUTH_SUCCESS',
    GITHUB_AUTH_FAIL = 'GITHUB_AUTH_FAIL',
    LOGOUT = 'LOGOUT',

    CHAT_ENGINE_LOGIN_SUCCESS = 'CHAT_ENGINE_LOGIN_SUCCESS',
    CHAT_ENGINE_LOGIN_FAIL = 'CHAT_ENGINE_LOGIN_FAIL',
    CHAT_ENGINE_SIGNUP_SUCCESS = 'CHAT_ENGINE_SIGNUP_SUCCESS',
    CHAT_ENGINE_SIGNUP_FAIL = 'CHAT_ENGINE_SIGNUP_FAIL',

    RESET_ERRORS = 'RESET_ERRORS',
}